scripts/data-collection/requirements.txt
playwright>=1.40.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

